refactor: 예약과 시간 domain 분
