+++ src/main/java/roomescape/domain/reservation/controller/AdminReservationController.java
@@ -1,16 +1,17 @@
 package roomescape.domain.reservation.controller;
 
-import java.net.URI;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.login.controller.MemberResolver;
 import roomescape.domain.member.domain.Member;
-import roomescape.domain.reservation.domain.reservation.Reservation;
+import roomescape.domain.reservation.domain.Reservation;
 import roomescape.domain.reservation.dto.ReservationAddRequest;
 import roomescape.domain.reservation.dto.ReservationResponse;
 import roomescape.domain.reservation.service.ReservationService;
+
+import java.net.URI;
 
 @RestController
 public class AdminReservationController {
