+++ src/main/java/roomescape/domain/reservation/controller/ReservationController.java
@@ -4,9 +4,14 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import roomescape.domain.login.controller.MemberResolver;
 import roomescape.domain.member.domain.Member;
-import roomescape.domain.reservation.domain.reservation.Reservation;
-import roomescape.domain.reservation.dto.*;
+import roomescape.domain.reservation.domain.Reservation;
+import roomescape.domain.reservation.dto.ReservationAddRequest;
+import roomescape.domain.reservation.dto.ReservationFindRequest;
+import roomescape.domain.reservation.dto.ReservationMineResponse;
+import roomescape.domain.reservation.dto.ReservationResponse;
 import roomescape.domain.reservation.service.ReservationService;
+import roomescape.domain.time.dto.BookableTimeResponse;
+import roomescape.domain.time.dto.BookableTimesRequest;
 
 import java.net.URI;
 import java.time.LocalDate;
