+++ src/main/java/roomescape/domain/reservation/service/ReservationService.java
@@ -4,16 +4,16 @@ import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
 import roomescape.domain.member.domain.Member;
 import roomescape.domain.member.repository.MemberRepository;
-import roomescape.domain.reservation.domain.reservation.Reservation;
-import roomescape.domain.reservation.domain.reservationTime.ReservationTime;
-import roomescape.domain.reservation.dto.BookableTimeResponse;
-import roomescape.domain.reservation.dto.BookableTimesRequest;
+import roomescape.domain.reservation.domain.Reservation;
 import roomescape.domain.reservation.dto.ReservationAddRequest;
 import roomescape.domain.reservation.dto.ReservationMineResponse;
 import roomescape.domain.reservation.repository.ReservationRepository;
-import roomescape.domain.reservation.repository.ReservationTimeRepository;
 import roomescape.domain.theme.domain.Theme;
 import roomescape.domain.theme.repository.ThemeRepository;
+import roomescape.domain.time.domain.ReservationTime;
+import roomescape.domain.time.dto.BookableTimeResponse;
+import roomescape.domain.time.dto.BookableTimesRequest;
+import roomescape.domain.time.repository.ReservationTimeRepository;
 import roomescape.global.exception.DataConflictException;
 
 import java.time.LocalDate;
