+++ src/main/java/roomescape/domain/time/dto/BookableTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.domain.reservation.dto;
+package roomescape.domain.time.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonProperty;
