+++ src/main/java/roomescape/domain/time/service/AdminReservationTimeService.java
@@ -1,10 +1,10 @@
-package roomescape.domain.reservation.service;
+package roomescape.domain.time.service;
 
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
-import roomescape.domain.reservation.domain.reservationTime.ReservationTime;
-import roomescape.domain.reservation.dto.ReservationTimeAddRequest;
-import roomescape.domain.reservation.repository.ReservationTimeRepository;
+import roomescape.domain.time.domain.ReservationTime;
+import roomescape.domain.time.dto.ReservationTimeAddRequest;
+import roomescape.domain.time.repository.ReservationTimeRepository;
 import roomescape.global.exception.DataConflictException;
 
 import java.util.List;
